+++ packages/react-core/src/constants/Routes.ts
@@ -57,6 +57,7 @@ export const EVENTS_SUBSCRIBED_ROUTE_NAME = 'events_subscribed';
 export const EVENTS_HIGHLIGHT_ROUTE_NAME = 'events_highlight';
 export const LIVESTREAM_ROUTE_NAME = 'live';
 export const CHECKOUT_PAYMENT = 'checkout_payment';
+export const LEADERBOARD_ROUTE_NAME = 'leaderboard';
 
 /**
  * Default Routes
@@ -116,4 +117,5 @@ export const defaultRoutes: {[k: string]: string} = {
   [COURSE_LESSON_EDIT_ROUTE_NAME]: '/course/:id/:slug/section/:section_id/lesson/:lesson_id/edit',
   [LIVESTREAM_ROUTE_NAME]: '/live/:slug/',
   [CHECKOUT_PAYMENT]: '/checkout/:content_type/:content_id/price/:price_id',
+  [LEADERBOARD_ROUTE_NAME]: '/leaderboard/',
 };