+++ packages/react-theme-default/src/index.ts
@@ -128,6 +128,9 @@ import SCLessonCommentActionsMenu from './components/SCLessonCommentActionsMenu'
 import SCLessonCommentObject from './components/SCLessonCommentObject';
 import SCLessonCommentObjects from './components/SCLessonCommentObjects';
 import SCCourseCompletedDialog from './components/SCCourseCompletedDialog';
+import SCLeaderboardInfoWidget from './components/SCLeaderboardInfoWidget';
+import SCLeaderboardPositionWidget from './components/SCLeaderboardPositionWidget';
+import SCLeaderboardWidget from './components/SCLeaderboardWidget';
 import SCLessonEditForm from './components/SCLessonEditForm';
 import SCLessonFilePreview from './components/SCLessonFilePreview';
 import SCLessonTemplate from './components/SCLessonTemplate';
@@ -414,6 +417,9 @@ const theme = {
     SCLessonCommentObject,
     SCLessonCommentObjects,
     SCCourseCompletedDialog,
+    SCLeaderboardInfoWidget,
+    SCLeaderboardPositionWidget,
+    SCLeaderboardWidget,
     SCLessonEditForm,
     SCLessonFilePreview,
     SCLessonTemplate,