+++ packages/react-ui/src/components/LeaderboardInfoWidget/index.tsx
@@ -1,0 +1,3 @@
+import LeaderboardInfoWidget, {LeaderboardInfoWidgetProps} from './LeaderboardInfoWidget';
+export default LeaderboardInfoWidget;
+export {LeaderboardInfoWidgetProps};