+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0-alpha.0](https://github.com/selfcommunity/community-js/compare/@selfcommunity/utils@1.0.19-alpha.0...@selfcommunity/utils@1.1.0-alpha.0) (2026-08-21)
+
+
+### Features
+
+* added getPeriodRange util and leaderboard period params ([14d8655](https://github.com/selfcommunity/community-js/commit/14d8655a49fdd7273f1977846422ea292c9e703e))
+
+
+
+
+
+## [1.0.19-alpha.0](https://github.com/selfcommunity/community-js/compare/@selfcommunity/utils@1.0.18...@selfcommunity/utils@1.0.19-alpha.0) (2026-08-20)
+
+**Note:** Version bump only for package @selfcommunity/utils
+
+
+
+
+
 ## [1.0.18](https://github.com/selfcommunity/community-js/compare/@selfcommunity/utils@1.0.18-alpha.1...@selfcommunity/utils@1.0.18) (2026-07-28)
 
 **Note:** Version bump only for package @selfcommunity/utils